change: Extract Col component from EntityPage module into own TS module
It's not entity related and a generic component for being used in
tables.
