+++ src/web/components/table/Col.tsx
@@ -1,0 +1,12 @@
+/* SPDX-FileCopyrightText: 2025 Greenbone AG
+ *
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ */
+
+import styled from 'styled-components';
+
+const Col = styled.col`
+  width: ${props => props.width};
+`;
+
+export default Col;
