+++ src/web/pages/certbund/Details.jsx
@@ -10,10 +10,10 @@ import SeverityBar from 'web/components/bar/SeverityBar';
 import Layout from 'web/components/layout/Layout';
 import ExternalLink from 'web/components/link/ExternalLink';
 import TableBody from 'web/components/table/Body';
+import Col from 'web/components/table/Col';
 import TableData from 'web/components/table/Data';
 import InfoTable from 'web/components/table/InfoTable';
 import TableRow from 'web/components/table/Row';
-import {Col} from 'web/entity/Page';
 import PropTypes from 'web/utils/PropTypes';
 
 const CertBundAdvDetails = ({entity}) => {
