+++ src/web/pages/overrides/Details.jsx
@@ -9,13 +9,13 @@ import React from 'react';
 import HorizontalSep from 'web/components/layout/HorizontalSep';
 import Layout from 'web/components/layout/Layout';
 import TableBody from 'web/components/table/Body';
+import Col from 'web/components/table/Col';
 import TableData from 'web/components/table/Data';
 import InfoTable from 'web/components/table/InfoTable';
 import TableRow from 'web/components/table/Row';
 import DetailsBlock from 'web/entity/Block';
 import EntityLink from 'web/entity/Link';
 import OverrideBox from 'web/entity/Override';
-import {Col} from 'web/entity/Page';
 import PropTypes from 'web/utils/PropTypes';
 import {
   translatedResultSeverityRiskFactor,
