+++ src/web/pages/permissions/Details.jsx
@@ -10,11 +10,11 @@ import React from 'react';
 import Divider from 'web/components/layout/Divider';
 import Layout from 'web/components/layout/Layout';
 import TableBody from 'web/components/table/Body';
+import Col from 'web/components/table/Col';
 import TableData from 'web/components/table/Data';
 import InfoTable from 'web/components/table/InfoTable';
 import TableRow from 'web/components/table/Row';
 import EntityLink from 'web/entity/Link';
-import {Col} from 'web/entity/Page';
 import PropTypes from 'web/utils/PropTypes';
 import {permissionDescription} from 'web/utils/Render';
