+++ src/web/pages/reports/details/Summary.jsx
@@ -14,10 +14,10 @@ import ErrorPanel from 'web/components/error/ErrorPanel';
 import Layout from 'web/components/layout/Layout';
 import DetailsLink from 'web/components/link/DetailsLink';
 import TableBody from 'web/components/table/Body';
+import Col from 'web/components/table/Col';
 import TableData from 'web/components/table/Data';
 import Table from 'web/components/table/InfoTable';
 import TableRow from 'web/components/table/Row';
-import {Col} from 'web/entity/Page';
 import PropTypes from 'web/utils/PropTypes';
 
 const UpdatingTable = styled(Table)`
