+++ src/web/pages/scanners/Details.jsx
@@ -11,11 +11,11 @@ import CertInfo from 'web/components/certinfo/CertInfo';
 import Layout from 'web/components/layout/Layout';
 import DetailsLink from 'web/components/link/DetailsLink';
 import TableBody from 'web/components/table/Body';
+import Col from 'web/components/table/Col';
 import TableData, {TableDataAlignTop} from 'web/components/table/Data';
 import InfoTable from 'web/components/table/InfoTable';
 import TableRow from 'web/components/table/Row';
 import DetailsBlock from 'web/entity/Block';
-import {Col} from 'web/entity/Page';
 import PropTypes from 'web/utils/PropTypes';
 
 const ScannerDetails = ({entity}) => {
