+++ src/web/pages/tags/Details.jsx
@@ -9,10 +9,10 @@ import {isDefined} from 'gmp/utils/identity';
 import React from 'react';
 import Layout from 'web/components/layout/Layout';
 import TableBody from 'web/components/table/Body';
+import Col from 'web/components/table/Col';
 import TableData from 'web/components/table/Data';
 import InfoTable from 'web/components/table/InfoTable';
 import TableRow from 'web/components/table/Row';
-import {Col} from 'web/entity/Page';
 import PropTypes from 'web/utils/PropTypes';
 import {renderYesNo} from 'web/utils/Render';
