+++ src/web/pages/tlscertificates/Details.jsx
@@ -10,10 +10,10 @@ import {connect} from 'react-redux';
 import DateTime from 'web/components/date/DateTime';
 import Layout from 'web/components/layout/Layout';
 import TableBody from 'web/components/table/Body';
+import Col from 'web/components/table/Col';
 import TableData from 'web/components/table/Data';
 import InfoTable from 'web/components/table/InfoTable';
 import TableRow from 'web/components/table/Row';
-import {Col} from 'web/entity/Page';
 import {
   loadEntity as loadTlsCertificate,
   selector as tlsCertificateSelector,
